+++ messages/en/AdminConsole.json
@@ -43,6 +43,8 @@
   "subTabSaAccessRequests": "JIT access requests",
   "subTabScim": "SCIM",
   "subTabDirectorySync": "Directory Sync",
+  "subTabMembersList": "List",
+  "subTabMembersTransferOwnership": "Transfer ownership",
 
   "sectionMembers": "Members",
   "sectionMembersDesc": "Manage members and roles.",
@@ -91,7 +93,5 @@
   "teamSectionAuditLogs": "Audit logs",
   "teamSectionAuditLogsDesc": "Review team activity and security events.",
 
-  "memberAddButton": "Add member",
-  "memberTransferOwnershipLink": "Transfer ownership",
-  "backToMembers": "Back to members"
+  "memberAddButton": "Add member"
 }
