+++ src/app/[locale]/admin/teams/[teamId]/members/transfer-ownership/page.tsx
@@ -53,7 +53,6 @@ export default function TeamTransferOwnershipPage({
 }) {
   const { teamId } = use(params);
   const t = useTranslations("Team");
-  const tAdmin = useTranslations("AdminConsole");
 
   const [team, setTeam] = useState<TeamInfo | null>(null);
   const [loadError, setLoadError] = useState(false);
@@ -144,14 +143,7 @@ export default function TeamTransferOwnershipPage({
   }
 
   return (
-    <div className="space-y-4">
-      <Link
-        href={`/admin/teams/${teamId}/members`}
-        className="inline-flex items-center text-sm text-muted-foreground hover:text-foreground transition-colors"
-      >
-        ← {tAdmin("backToMembers")}
-      </Link>
-      <Card>
+    <Card>
       <SectionCardHeader icon={Crown} title={t("transferTitle")} description={t("transferDescription")} />
       <CardContent className="space-y-4">
         {members.length > 1 && (
@@ -213,6 +205,5 @@ export default function TeamTransferOwnershipPage({
         )}
       </CardContent>
     </Card>
-    </div>
   );
 }
